% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misfit_epmfd.R
\name{misfit_epmfd}
\alias{misfit_epmfd}
\title{Compute person-fit statistics (polytomous)}
\usage{
misfit_epmfd(object, stats = c("auto", "lpz", "Gp", "U3p"), alpha = 0.05)
}
\arguments{
\item{object}{An \code{epmfd_scaled} object.}

\item{stats}{Character vector: any of \code{"auto"}, \code{"lpz"},
\code{"Gp"} , \code{"U3p"}.}

\item{alpha}{Significance level (default 0.05).}
}
\value{
An \code{epmfd_misfit} object.
}
\description{
Compute person-fit statistics (polytomous)
}
