% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{sampledata}
\alias{sampledata}
\title{Example Polytomous Response Data}
\format{
A data frame with 20 persons (rows) and 6 items (columns).
Each item takes ordered values 1–5.
}
\usage{
sampledata
}
\description{
A small toy dataset included in the \strong{epmfd} package, containing
polytomous item responses from simulated persons.
}
\examples{
data(sampledata)
head(sampledata)

}
\keyword{datasets}
