% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_epmfd.R
\name{summary.epmfd_clean}
\alias{summary.epmfd_clean}
\title{Summary method for \code{epmfd_clean} objects}
\usage{
\method{summary}{epmfd_clean}(object, ...)
}
\arguments{
\item{object}{An object of class \code{epmfd_clean}.}

\item{...}{Further arguments (ignored).}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  Invisibly returns a named list with summary numbers.
}\if{html}{\out{</div>}}
}
\description{
Summary method for \code{epmfd_clean} objects
}
