% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_epmfd.R
\name{clean_epmfd}
\alias{clean_epmfd}
\title{Remove mis-fitting persons}
\usage{
clean_epmfd(misfit, criterion = c("union", "intersection"))
}
\arguments{
\item{misfit}{An object returned by \code{\link[=misfit_epmfd]{misfit_epmfd()}}.}

\item{criterion}{"union" (default) = at least one statistic flags
a person; "intersection" = all statistics flag the person.}
}
\value{
An \code{epmfd_clean} object that also contains a ready-to-use
\code{epmfd_raw} object in the slot \code{raw}, so you can call
\code{scale_epmfd(clean$raw, ...)} directly.
}
\description{
Remove mis-fitting persons
}
