% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_epmfd.R
\name{load_epmfd}
\alias{load_epmfd}
\title{Load and validate data for epmfd workflow}
\usage{
load_epmfd(data, id_col = NULL, likert_levels = NULL)
}
\arguments{
\item{data}{A data.frame / tibble: rows = persons, columns = items.}

\item{id_col}{Optional \if{html}{\out{<character>}}: column name holding unique IDs.}

\item{likert_levels}{Optional \if{html}{\out{<integer>}}: maximum category K; if NULL
it is inferred from the data.}
}
\value{
An object of class \code{epmfd_raw}.
}
\description{
Load and validate data for epmfd workflow
}
