% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misfit_epmfd.R
\name{misfit_epmfd}
\alias{misfit_epmfd}
\title{Compute Person Misfit Statistics}
\usage{
misfit_epmfd(
  object,
  stats = c("auto", "lpz", "Gp", "Gpn", "U3p"),
  alpha = 0.05
)
}
\arguments{
\item{object}{An \code{epmfd_scaled}.}

\item{stats}{"auto", "lpz", "Gp", "Gpn", "U3p".}

\item{alpha}{Significance level (0.05).}
}
\value{
An \code{epmfd_misfit} object with misfit flags.
}
\description{
Compute Person Misfit Statistics
}
