% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_misfit_scatter.R
\name{plot_misfit_scatter}
\alias{plot_misfit_scatter}
\title{Scatter / bubble plot for epmfd_misfit}
\usage{
plot_misfit_scatter(
  object,
  x_stat = NULL,
  y_stat = NULL,
  z_stat = NULL,
  bubble = TRUE,
  alpha = 0.05,
  x_cut = NULL,
  y_cut = NULL,
  z_cut = NULL,
  label_ids = FALSE,
  ...
)
}
\arguments{
\item{x_stat, y_stat, z_stat}{Character names of statistics in
\code{misfit$scores}.  If all three are \code{NULL}, the first three
available statistics are chosen automatically.}
}
\description{
Scatter / bubble plot for epmfd_misfit
}
