% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_epmfd.R
\name{print.epmfd_misfit}
\alias{print.epmfd_misfit}
\title{Print method for \code{epmfd_misfit} objects}
\usage{
\method{print}{epmfd_misfit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{epmfd_misfit}.}

\item{...}{Further arguments passed to \code{\link[utils:head]{utils::head()}}.}
}
\description{
Print method for \code{epmfd_misfit} objects
}
