% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_epmfd.R
\name{scale_epmfd}
\alias{scale_epmfd}
\title{Scale Data via MIRT or Mokken}
\usage{
scale_epmfd(
  object,
  method = c("auto", "mirt", "mokken"),
  H_thr = 0.3,
  a_thr = 0.5,
  ...
)
}
\arguments{
\item{object}{An \code{epmfd_raw} created by \code{\link[=load_epmfd]{load_epmfd()}}.}

\item{method}{"auto", "mirt", or "mokken".}

\item{H_thr}{Numeric. Min scalability coefficient for Mokken (default 0.3).}

\item{a_thr}{Numeric. Min discrimination for MIRT (default 0.5).}

\item{...}{Further arguments passed to \code{mirt()} or \code{mokken::check()}/\code{mokken::aisp}.}
}
\value{
An object of class \code{epmfd_scaled}.
}
\description{
Scale Data via MIRT or Mokken
}
